+++ CHANGELOG.md
@@ -12,7 +12,7 @@ This is where we capture breaking changes
 
 - Removed size properties from the .box class on search results (add bootstrap
   col classes instead)
-- Removed the "enableGeotagging" config parameter, use "features.geotagging"
+- Removed the "enableGeotagging" config parameter, use "features.geoTagging"
   instead.
 - Moved favicons (apart from favicon.ico) from root to images/favicons